% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_occupancy.r
\name{bed_occupancy}
\alias{bed_occupancy}
\title{Plot bed occupancy in Slovakia}
\usage{
bed_occupancy()
}
\value{
plot
}
\description{
To update the data, first run "data-raw/scripts/hospitalisations.r" and
update, e.g., with \code{devtools::load_all}
}
