% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_table.r
\name{county_table}
\alias{county_table}
\title{Produce a table of prevalence ratios in counties}
\usage{
county_table(file = NULL)
}
\arguments{
\item{file}{file name to save the table to (pdf)}
}
\value{
NULL (invisibly)
}
\description{
Produce a table of prevalence ratios in counties
}
