% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_min_specificity.r
\name{estimate_min_specificity}
\alias{estimate_min_specificity}
\title{Estimate minimum specificity of the antigen test used for mass testing in
Slovakia}
\usage{
estimate_min_specificity(cutoff = 0.95)
}
\arguments{
\item{cutoff}{probability cut-off, i.e. x if minimal specificty is to be
estabilshed with probability >= x}
}
\description{
Estimate minimum specificity of the antigen test used for mass testing in
Slovakia
}
