% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.r
\name{make_labels}
\alias{make_labels}
\title{Helper function for lables in plot}
\usage{
make_labels(labels)
}
\arguments{
\item{labels}{lables}
}
\value{
a list of split labels
}
\description{
Helper function for lables in plot
}
\keyword{internal}
