% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_incidence.r
\name{pcr_incidence}
\alias{pcr_incidence}
\title{Plot PCR positive incidence and mass testing timings for Slovakia}
\usage{
pcr_incidence()
}
\value{
a plot with PCR positive incidence and mass testing timings
}
\description{
Plot PCR positive incidence and mass testing timings for Slovakia
}
