% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predicted_prevalence.r
\name{plot_predicted_prevalence}
\alias{plot_predicted_prevalence}
\title{Plot infection prevalence as predicted by EpiNow2}
\usage{
plot_predicted_prevalence(counties = NULL)
}
\arguments{
\item{counties}{counties to plot, including the possibilities of "pilot" (for all pilot counties), "non-pilot" (for all non-pilot counties) and "all"}
}
\value{
ggplot object
}
\description{
Plot infection prevalence as predicted by EpiNow2
}
