% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.r
\name{prevalence}
\alias{prevalence}
\title{Plot prevalence with confidence intervals as measured in national mass-testing in Slovakia}
\usage{
prevalence()
}
\value{
plot of prevalence with exact binomial confidence intervals
}
\description{
Plot prevalence with confidence intervals as measured in national mass-testing in Slovakia
}
