% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{regression}
\alias{regression}
\title{Regression analysis of prevalence reduction}
\usage{
regression(file = NULL)
}
\arguments{
\item{file}{a file name to save the table of covariates to, if desired}
}
\value{
the result of a \code{glm} call
}
\description{
Regression analysis of prevalence reduction
}
