% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratios.r
\name{risk_ratios}
\alias{risk_ratios}
\title{Calculate risk ratios for mass testing in Slovakia}
\usage{
risk_ratios()
}
\value{
alist of two lists: "figures" (containing Fig. 2a and 2b in the manuscript) and "tables" (containing the total prevalence risk ratios in pilot regions ("pilot") and those regions that had two rounds of testing ("nation")
}
\description{
Calculate risk ratios for mass testing in Slovakia
}
