% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_ratios.r
\name{self_adjusted_risk_ratios}
\alias{self_adjusted_risk_ratios}
\title{Self-adjusted risk ratios}
\usage{
self_adjusted_risk_ratios()
}
\value{
a list of risk ratios by county, pilot vs. non-pilot, and overall
}
\description{
Self-adjusted risk ratios
}
