% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_names.r
\name{simplify_names}
\alias{simplify_names}
\title{Sipmlify county names}
\usage{
simplify_names(x)
}
\arguments{
\item{x}{data frame with a "county" column that contains the original county name}
}
\value{
data frame with a "simple_name" column that contains the simplified name
}
\description{
Internal function to match county names for incidence data wiht county names for esting data
}
\author{
Sebastian Funk
}
\keyword{internal}
