% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{slovakia_shape}
\alias{slovakia_shape}
\title{Shapes}
\description{
Polygons for counties in Slovakia. Available from the National Land Registry
at https://www.geoportal.sk/sk/zbgis_smd/na-stiahnutie/
}
