% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{covariates}
\alias{covariates}
\title{Covariates}
\description{
Covariates for prevalence: unemployment rate, population in the Marginalised
Roma Community, mean age and population density.
}
