% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.r
\name{regression}
\alias{regression}
\title{Regression analysis of prevalence}
\usage{
regression(table_file = NULL)
}
\arguments{
\item{table_file}{(optional) file to save parameter table to}
}
\value{
a list of two plots (effect size and posterior predicitons) and a
table of effect size estimates
}
\description{
Regression analysis of prevalence
}
